Update AI to handle new bonus-tile draw mechanic in v1.2.
The new high-tile draw system presents a list of up to three possible
bonus cards. We deal with this by passing around tilesets (sets of
candidate next tiles).

The OCR module has been revamped and improved to deal with this change,
as well as layout and appearance changes from the recent patches.